refactor(root): split root.go into multiple files
close #463
