+++ mgc/cli/cmd/get_next_unknown_cmd.go
@@ -1,0 +1,75 @@
+package cmd
+
+import (
+	"strings"
+
+	"github.com/spf13/cobra"
+	flag "github.com/spf13/pflag"
+)
+
+// BEGIN: copied from cobra: command.go
+
+// These are required to get the next unknown command and its arguments without the flags
+
+func hasNoOptDefVal(name string, fs *flag.FlagSet) bool {
+	flag := fs.Lookup(name)
+	if flag == nil {
+		return false
+	}
+	return flag.NoOptDefVal != ""
+}
+
+func shortHasNoOptDefVal(name string, fs *flag.FlagSet) bool {
+	if len(name) == 0 {
+		return false
+	}
+
+	flag := fs.ShorthandLookup(name[:1])
+	if flag == nil {
+		return false
+	}
+	return flag.NoOptDefVal != ""
+}
+
+func isFlagArg(arg string) bool {
+	return ((len(arg) >= 3 && arg[0:2] == "--") ||
+		(len(arg) >= 2 && arg[0] == '-' && arg[1] != '-'))
+}
+
+// END: copied from cobra: command.go
+
+// this is a modified Traverse() from cobra: command.go
+// we need it to get the next command skipping flags
+// it's far from ideal, for instance this is a bug:
+//
+//	$ main-cmd sub-cmd -h sub-sub-cmd
+//
+// it will take `sub-sub-cmd` as -h argument (but it's not a boolean)
+func getNextUnknownCommand(c *cobra.Command, args []string) (*string, []string) {
+	inFlag := false
+
+	for i, arg := range args {
+		switch {
+		// A long flag with a space separated value
+		case strings.HasPrefix(arg, "--") && !strings.Contains(arg, "="):
+			// TODO: this isn't quite right, we should really check ahead for 'true' or 'false'
+			inFlag = !hasNoOptDefVal(arg[2:], c.Flags())
+			continue
+		// A short flag with a space separated value
+		case strings.HasPrefix(arg, "-") && !strings.Contains(arg, "=") && len(arg) == 2 && !shortHasNoOptDefVal(arg[1:], c.Flags()):
+			inFlag = true
+			continue
+		// The value for a flag
+		case inFlag:
+			inFlag = false
+			continue
+		// A flag without a value, or with an `=` separated value
+		case isFlagArg(arg):
+			continue
+		}
+
+		return &arg, args[i+1:]
+	}
+
+	return nil, args
+}
